Renamed fixture heirarchy
