+++ cubano-concordion/src/main/java/org/concordion/cubano/framework/ConcordionFixture.java
@@ -1,15 +1,16 @@
-package demo;
+package org.concordion.cubano.framework;
 
 import org.concordion.api.extension.Extension;
-import org.concordion.cubano.framework.ConcordionBase;
 import org.concordion.ext.LoggingFormatterExtension;
 import org.concordion.ext.StoryboardExtension;
 import org.concordion.ext.StoryboardLogListener;
+import org.concordion.logback.LogbackAdaptor;
 
 /**
- * Project customisation layer.
- */
-public class ConcordionTempBase extends ConcordionBase {
+ * Concordion fixture for inheritance by any test classes.
+ * Includes and configures the Storyboard and Logging extensions.
+ **/
+public abstract class ConcordionFixture extends ConcordionBase {
     @Extension
     private final StoryboardExtension storyboard = new StoryboardExtension();
 
@@ -17,10 +18,11 @@ public class ConcordionTempBase extends ConcordionBase {
     private final LoggingFormatterExtension loggerExtension = new LoggingFormatterExtension()
             .registerListener(new StoryboardLogListener(getStoryboard()));
 
-    /**
-     * @return A reference to the Storyboard extension.
-     */
     protected StoryboardExtension getStoryboard() {
         return storyboard;
+    }
+
+    static {
+        LogbackAdaptor.logInternalStatus();
     }
 }
